+++ nts.uk/web/nts.uk.pr.web/src/main/webapp/view/qmm/006/a/viewmodel.ts
@@ -9,13 +9,23 @@ module qmm006.a.viewmodel {
 
         currentLineBank: KnockoutObservable<LineBank>;
         roundingRules: KnockoutObservableArray<any>;
-
+        //trang thai cua nut delete
+        isDeleteEnable: KnockoutObservable<boolean>;
+        //trang thai cua INP_001
         isEnable: KnockoutObservable<boolean>;
+        //xuat hien dau "-" (LBL_006) hay khong
         isAppear: KnockoutObservable<boolean>;
 
         bankName: KnockoutObservable<string>;
         branchName: KnockoutObservable<string>;
         messageList: KnockoutObservableArray<any>;
+
+        countItems: KnockoutObservable<number>;
+        findIndex: KnockoutObservable<number>;
+        //chi cho phep select row dau tien trong lan findAll() dau tien
+        isSelectdFirstRow: KnockoutObservable<boolean>;
+
+        dirty: nts.uk.ui.DirtyChecker;
 
         constructor() {
             var self = this;
@@ -28,14 +38,14 @@ module qmm006.a.viewmodel {
                 { messageId: "ER001", message: "＊が入力されていません。" },
                 { messageId: "ER007", message: "＊が選択されていません。" },
                 { messageId: "ER005", message: "入力した＊は既に存在しています。\r\n ＊を確認してください。" },
-                { messageId: "ER008", message: "選択された＊は使用されているため削除できません。" }
+                { messageId: "ER008", message: "選択された＊は使用されているため削除できません。" },
             ]);
 
             self.columns = ko.observableArray([
-                { headerText: 'コード', key: 'lineBankCode', width: 45 },
-                { headerText: '名称', key: 'lineBankName', width: 120 },
-                { headerText: '口座区分', key: 'accountAtr', width: 110 },
-                { headerText: '口座番号', key: 'accountNo', width: 100 }
+                { headerText: 'コード', key: 'lineBankCode', width: 45, formatter: _.escape },
+                { headerText: '名称', key: 'lineBankName', width: 120, formatter: _.escape },
+                { headerText: '口座区分', key: 'accountAtr', width: 110, formatter: _.escape },
+                { headerText: '口座番号', key: 'accountNo', width: 100, formatter: _.escape }
             ]);
             self.currentLineBank = ko.observable(new LineBank(null, null, null, null, 0, null, null, null, []));
             self.roundingRules = ko.observableArray([
@@ -45,32 +55,13 @@ module qmm006.a.viewmodel {
 
             self.bankName = ko.observable('');
             self.branchName = ko.observable('');
+            self.findIndex = ko.observable(0);
+            self.countItems = ko.observable(0);
 
+            self.isDeleteEnable = ko.observable(true);
             self.isEnable = ko.observable(false);
             self.isAppear = ko.observable(true);
-
-            //theo doi su thay doi cua currentCode khi click
-            self.currentCode.subscribe(function(codeChanged) {
-                var lineBank = self.findLineBank(codeChanged);
-                var tmp = _.find(self.dataSource2(), function(x) {
-                    return x.code === lineBank.bankCode();
-                });
-                var tmp1 = _.find(self.dataSource2(), function(x) {
-                    return x.code === lineBank.branchCode() && x.parentCode === lineBank.bankCode();
-                });
-                if (tmp != undefined) {
-                    self.bankName(tmp.name);
-                } else self.bankName('');
-                if (tmp1 != undefined) {
-                    self.branchName(tmp1.name);
-                    self.isAppear(true);
-                } else {
-                    self.isAppear(false);
-                    self.branchName('');
-                }
-                self.isEnable(false);
-                self.currentLineBank(lineBank);
-            });
+            self.isSelectdFirstRow = ko.observable(true);
         }
 
         startPage() {
@@ -91,9 +82,7 @@ module qmm006.a.viewmodel {
         OpenBDialog() {
             var self = this;
             var lineBank = self.currentLineBank();
-            if (lineBank.bankCode() != null && lineBank.branchCode != null) {
-                nts.uk.ui.windows.setShared("branchCode", lineBank.bankCode() + lineBank.branchCode(), true);
-            }
+            nts.uk.ui.windows.setShared("bankBranchCode", lineBank.bankCode() + lineBank.branchCode(), true);
             nts.uk.ui.windows.sub.modal("/view/qmm/006/b/index.xhtml", { title: "銀行情報一覧", dialogClass: "no-close" }).onClosed(function() {
                 if (nts.uk.ui.windows.getShared("selectedBank") != null) {
                     if (nts.uk.ui.windows.getShared("selectedBank").parentCode == null) {
@@ -115,7 +104,36 @@ module qmm006.a.viewmodel {
 
         OpenCDialog() {
             var self = this;
-            nts.uk.ui.windows.sub.modal("/view/qmm/006/c/index.xhtml", { title: "振込元銀行の登録　＞　振込元銀行", dialogClass: "no-close" });
+            if (self.dirty.isDirty()) {
+                nts.uk.ui.dialog.confirm("変更された内容が登録されていません。\r\n よろしいですか。").ifYes(function() {
+                    nts.uk.ui.windows.sub.modal("/view/qmm/006/c/index.xhtml", { title: "振込元銀行の登録　＞　振込元銀行", dialogClass: "no-close" }).onClosed(function() {
+                        self.findAll().done(function() {
+                            if (nts.uk.ui.windows.getShared("currentCode")) {
+                                self.currentCode(nts.uk.ui.windows.getShared("currentCode"));
+                            }
+                        });
+                    });
+                }).ifNo(function() {
+                })
+            }
+        }
+
+        BTN007() {
+            var self = this;
+            if (self.dirty.isDirty()) {
+                nts.uk.ui.dialog.confirm("変更された内容が登録されていません。\r\n よろしいですか。").ifYes(function() {
+                }).ifNo(function() {
+                })
+            }
+        }
+
+        BTN005() {
+            var self = this;
+            if (self.dirty.isDirty()) {
+                nts.uk.ui.dialog.confirm("変更された内容が登録されていません。\r\n よろしいですか。").ifYes(function() {
+                }).ifNo(function() {
+                })
+            }
         }
 
         findLineBank(codeNew): LineBank {
@@ -131,12 +149,18 @@ module qmm006.a.viewmodel {
 
         clearForm(): any {
             var self = this;
-            self.currentLineBank(new LineBank(null, null, null, null, 0, null, null, null, []));
-            self.currentCode("");
-            self.isEnable(true);
+            if (self.dirty.isDirty()) {
+                nts.uk.ui.dialog.confirm("変更された内容が登録されていません。\r\n よろしいですか。").ifYes(function() {
+                    self.currentLineBank(new LineBank(null, null, null, null, 0, null, null, null, []));
+                    self.currentCode("");
+                    self.isEnable(true);
+                    self.isDeleteEnable(false);
+                }).ifNo(function() {
+                })
+            }
         }
 
-        selectedFirstItem(item): LineBank {
+        selectedItem(item): LineBank {
             var self = this;
             self.currentCode(item.lineBankCode);
             return new LineBank(item.bankCode, item.branchCode,
@@ -175,13 +199,78 @@ module qmm006.a.viewmodel {
             var dfd = $.Deferred();
             qmm006.a.service.findAll()
                 .done(function(data) {
+                    self.countItems(data.length);
                     if (data.length > 0) {
+                        if (self.isSelectdFirstRow()) {
+                            //load lai list va chi vao row dau tien
+                            self.currentLineBank(self.selectedItem(data[0]));
+                            
+                            //set gia tri cho LBL_005 va LBL_007
+                            var tmp = _.find(self.dataSource2(), function(x) {
+                                return x.code === self.currentLineBank().bankCode();
+                            });
+                            var tmp1 = _.find(self.dataSource2(), function(x) {
+                                return x.code === self.currentLineBank().branchCode() && x.parentCode === self.currentLineBank().bankCode();
+                            });
+                            if (tmp != undefined) {
+                                self.bankName(tmp.name);
+                            } else {
+                                self.bankName('');
+                            }
+                            if (tmp1 != undefined) {
+                                self.branchName(tmp1.name);
+                                self.isAppear(true);
+                            } else {
+                                self.isAppear(false);
+                                self.branchName('');
+                            }
+                            //
+                            self.dirty = new nts.uk.ui.DirtyChecker(self.currentLineBank);
+                            self.findIndex(0);
+                            self.isSelectdFirstRow(false);
+                        }
                         self.items(data);
-                        self.currentLineBank(self.selectedFirstItem(data[0]));
+                        self.isDeleteEnable(true);
+                        
+                        //theo doi su thay doi cua currentCode khi click
+                        self.currentCode.subscribe(function(codeChanged) {
+                            self.findIndex(
+                                _.findIndex(data, function(x) {
+                                    return x.lineBankCode == codeChanged;
+                                }));
+                            self.clearError();
+                            self.isDeleteEnable(true);
+                            var lineBank = self.findLineBank(codeChanged);
+                            //set gia tri cho LBL_005 va LBL_007
+                            var tmp = _.find(self.dataSource2(), function(x) {
+                                return x.code === lineBank.bankCode();
+                            });
+                            var tmp1 = _.find(self.dataSource2(), function(x) {
+                                return x.code === lineBank.branchCode() && x.parentCode === lineBank.bankCode();
+                            });
+                            if (tmp != undefined) {
+                                self.bankName(tmp.name);
+                            } else {
+                                self.bankName('');
+                            }
+                            if (tmp1 != undefined) {
+                                self.branchName(tmp1.name);
+                                self.isAppear(true);
+                            } else {
+                                self.isAppear(false);
+                                self.branchName('');
+                            }
+                            //
+                            self.isEnable(false);
+                            self.currentLineBank(lineBank);
+                            self.dirty = new nts.uk.ui.DirtyChecker(self.currentLineBank);
+                        });
+
                     } else {
                         self.items([]);
                         self.clearForm();
                         self.isAppear(false);
+                        self.isDeleteEnable(false);
                     }
                     dfd.resolve();
                 }).fail(function(res) {
@@ -217,6 +306,11 @@ module qmm006.a.viewmodel {
             };
             qmm006.a.service.saveData(self.isEnable(), command)
                 .done(function() {
+                    //load lai list va chi vao row moi them
+                    $.when(self.findAll()).done(function() {
+                        self.currentCode(command.lineBankCode);
+                        self.isEnable(false);
+                    });
                 }).fail(function(error) {
                     if (error.messageId == self.messageList()[0].messageId) {
                         var message = self.messageList()[0].message;
@@ -241,46 +335,50 @@ module qmm006.a.viewmodel {
                         var message = self.messageList()[2].message;
                         $('#A_INP_001').ntsError('set', message);
                     }
-                    console.log(error);
-                }).then(function() {
-                    //load lai list va chi vao row moi them
-                    $.when(self.findAll()).done(function() {
-                        self.currentCode(command.lineBankCode);
-                        self.isEnable(false);
-                    });
                 });
             return dfd.promise();
         }
 
+
         remove() {
             var self = this;
-            if (self.items().length == 0) {
-                return;
-            }
+            nts.uk.ui.dialog.confirm("データを削除します。\r\nよろしいですか？").ifYes(function() {
+                var command = {
+                    accountAtr: self.currentLineBank().accountAtr(),
+                    bankCode: self.currentLineBank().bankCode(),
+                    branchCode: self.currentLineBank().branchCode(),
+                    consignor: [
+                        { "code": self.currentLineBank().consignors()[0].consignorCode(), "memo": self.currentLineBank().consignors()[0].consignorMemo() },
+                        { "code": self.currentLineBank().consignors()[1].consignorCode(), "memo": self.currentLineBank().consignors()[1].consignorMemo() },
+                        { "code": self.currentLineBank().consignors()[2].consignorCode(), "memo": self.currentLineBank().consignors()[2].consignorMemo() },
+                        { "code": self.currentLineBank().consignors()[3].consignorCode(), "memo": self.currentLineBank().consignors()[3].consignorMemo() },
+                        { "code": self.currentLineBank().consignors()[4].consignorCode(), "memo": self.currentLineBank().consignors()[4].consignorMemo() },
+                    ],
+                    lineBankCode: self.currentLineBank().lineBankCode(),
+                    lineBankName: self.currentLineBank().lineBankName(),
+                    memo: self.currentLineBank().memo(),
+                    requesterName: self.currentLineBank().requesterName()
+                };
+                qmm006.a.service.remove(command)
+                    .done(function() {
+                        self.findAll().done(function() {
+                            var data = self.items();
+                            if (self.countItems() == self.findIndex()) {
+                                self.currentCode((data[self.findIndex() - 1]).lineBankCode);
+                            }
+                            else if (self.findIndex() == -1) {
+                                self.clearForm();
+                            } else {
+                                self.currentCode((data[self.findIndex()]).lineBankCode);
+                            }
+                        })
+                    }).fail(function() {
+                        nts.uk.ui.dialog.alert(self.messageList()[3].message);
+                    });
 
-            var command = {
-                accountAtr: self.currentLineBank().accountAtr(),
-                accountNo: self.currentLineBank().accountNo(),
-                bankCode: self.currentLineBank().bankCode(),
-                branchCode: self.currentLineBank().branchCode(),
-                consignor: [
-                    { "code": self.currentLineBank().consignors()[0].consignorCode(), "memo": self.currentLineBank().consignors()[0].consignorMemo() },
-                    { "code": self.currentLineBank().consignors()[1].consignorCode(), "memo": self.currentLineBank().consignors()[1].consignorMemo() },
-                    { "code": self.currentLineBank().consignors()[2].consignorCode(), "memo": self.currentLineBank().consignors()[2].consignorMemo() },
-                    { "code": self.currentLineBank().consignors()[3].consignorCode(), "memo": self.currentLineBank().consignors()[3].consignorMemo() },
-                    { "code": self.currentLineBank().consignors()[4].consignorCode(), "memo": self.currentLineBank().consignors()[4].consignorMemo() },
-                ],
-                lineBankCode: self.currentLineBank().lineBankCode(),
-                lineBankName: self.currentLineBank().lineBankName(),
-                memo: self.currentLineBank().memo(),
-                requesterName: self.currentLineBank().requesterName()
-            };
-            qmm006.a.service.remove(command)
-                .done(function() {
-                    self.findAll();
-                }).fail(function() {
-                    alert(self.messageList()[3].message);
-                });
+            }).ifNo(function() {
+                return;
+            });
         }
 
         clearError() {
@@ -292,6 +390,7 @@ module qmm006.a.viewmodel {
         }
     }
 
+
     class LineBank {
         bankCode: KnockoutObservable<string>;
         branchCode: KnockoutObservable<string>;
@@ -302,7 +401,6 @@ module qmm006.a.viewmodel {
         accountNo: KnockoutObservable<string>;
         memo: KnockoutObservable<string>;
         requesterName: KnockoutObservable<string>;
-
 
         constructor(bankCode: string, branchCode: string, lineBankCode: string, lineBankName: string, accountAtr: number, accountNo: string, memo: string, requesterName: string, consignors: Array<any>) {
             this.bankCode = ko.observable(bankCode);
@@ -358,6 +456,7 @@ module qmm006.a.viewmodel {
                         break;
                 }
             }
+
         }
     }
